完成角色管理、人员管理页面和部分后台

+++ ubtms/src/main/webapp/resources/js/school/schoolMng.js
@@ -1,26 +1,27 @@
-﻿
-
-var schoolMng={
-    serachClick:function () {
+﻿var schoolMng = {
+    serachClick: function () {
         $("#tb_schools").bootstrapTable('refresh');
     },
 
-    URL:{
-        getSChools:function () {
+    URL: {
+        getRoles: function () {
             return 'school/schoolGetAction';
         },
-        add: function() {
+        add: function () {
             return 'school/schoolAddPage';
         },
-        detailAndEdit:function(){
+        detailAndEdit: function () {
             return '/school/schoolViewAndEditAction';
         },
-        delete:function () {
+        delete: function () {
             return "/school/schoolDelAction";
+        },
+        changeState: function () {
+            return "/school/schoolStateChangeAction";
         }
     },
 
-    init:function () {
+    init: function () {
         //1.初始化Table
         var oTable = new TableInit();
         oTable.Init();
@@ -34,11 +35,11 @@ var schoolMng={
 var TableInit = function () {
     //debugger;
     var oTableInit = new Object();
-    oTableInit.curPageNum=0;
+    oTableInit.curPageNum = 0;
     //初始化Table
     oTableInit.Init = function () {
         $('#tb_schools').bootstrapTable({
-            url: schoolMng.URL.getSChools(),         //请求后台的URL（*）
+            url: schoolMng.URL.getRoles(),         //请求后台的URL（*）
             method: 'get',                      //请求方式（*）
             toolbar: '#toolbar',                //工具按钮用哪个容器
             striped: true,                      //是否显示行间隔色
@@ -48,73 +49,72 @@ var TableInit = function () {
             sortOrder: "asc",                   //排序方式
             queryParams: oTableInit.queryParams,//传递参数（*）
             sidePagination: "server",           //分页方式：client客户端分页，server服务端分页（*）
-            pageNumber:1,                       //初始化加载第一页，默认第一页
+            pageNumber: 1,                       //初始化加载第一页，默认第一页
             pageSize: 10,                       //每页的记录行数（*）
-            pageList: [10,20],        //可供选择的每页的行数（*）
+            pageList: [10, 20],        //可供选择的每页的行数（*）
             search: false,                       //是否显示表格搜索，此搜索是客户端搜索，不会进服务端，所以，个人感觉意义不大
             strictSearch: true,
             showColumns: false,                  //是否显示列筛选按钮
             showRefresh: false,                  //是否显示刷新按钮
             minimumCountColumns: 2,             //最少允许的列数
             clickToSelect: true,                //是否启用点击选中行
-            height: 500,                        //行高，如果没有设置height属性，表格自动根据记录条数觉得表格高度
             uniqueId: "ID",                     //每一行的唯一标识，一般为主键列
-            showToggle:false,                    //是否显示详细视图和列表视图的切换按钮
+            showToggle: false,                    //是否显示详细视图和列表视图的切换按钮
             cardView: false,                    //是否显示详细视图
             detailView: false,                   //是否显示父子表
             columns: [{
-                title:null,
+                title: null,
                 checkbox: true,
-                width:'4%'
+                width: '4%'
             }, {
                 align: 'center',
                 title: '序号',
-                formatter:function(value,row,index){
-                   return index+1+oTableInit.curPageNum;
+                formatter: function (value, row, index) {
+                    return index + 1 + oTableInit.curPageNum;
                 },
-                width:'4%'
-            },{
+                width: '4%'
+            }, {
                 align: 'center',
                 field: 'schName',
                 title: '校名',
-                width:'52%'
-            },{
+                width: '52%'
+            }, {
                 align: 'center',
                 title: '状态',
-                formatter:function(value,row,index){
-                    if(row.state==1)
+                formatter: function (value, row, index) {
+                    if (row.state == 1)
                         return "正常";
                     else
                         return "禁用";
                 },
-                width:'20%'
+                width: '20%'
             }, {
                 align: 'center',
                 title: '操作',
-                formatter:function(value,row,index,params){
+                formatter: function (value, row, index, params) {
                     var editState = $('#schoolEditP').val();
-                    var detail = "<a href="+schoolMng.URL.detailAndEdit()+"?schId="+row.schId+"&type=0><i class='glyphicon glyphicon-eye-open'></i>&nbsp;查看</a>";
-                    var edit = "<a href="+schoolMng.URL.detailAndEdit()+"?schId="+row.schId+"&type=1 style='margin-left: 30px'><i class='glyphicon glyphicon-pencil'></i>&nbsp;编辑</a>";
-                    if(editState==1){
-                        return detail+edit;
-                    }else {
+                    var detail = "<a href=" + schoolMng.URL.detailAndEdit() + "?schId=" + row.schId + "&type=0><i class='glyphicon glyphicon-eye-open'></i>&nbsp;查看</a>";
+                    var edit = "<a href=" + schoolMng.URL.detailAndEdit() + "?schId=" + row.schId + "&type=1 style='margin-left: 30px'><i class='glyphicon glyphicon-pencil'></i>&nbsp;编辑</a>";
+                    if (editState == 1) {
+                        return detail + edit;
+                    } else {
                         return detail;
                     }
                 },
-                width:'20%'
+                width: '20%'
             },]
         });
     };
-    
+
     //得到查询的参数
     oTableInit.queryParams = function (params) {
-        oTableInit.curPageNum=params.offset;
-        var schoolState=$('#searchSchoolSate').val();
+        oTableInit.curPageNum = params.offset;
+        var schoolState = $('#searchSchoolSate').val();
         var temp = {   //这里的键的名字和控制器的变量名必须一直，这边改动，控制器也需要改成一样的
             limit: params.limit,   //页面大小
             offset: params.offset,  //页码
-            schoolName:$('#searchSchoolName').val(),
-            state:schoolState
+            schoolName: $('#searchSchoolName').val(),
+            state: schoolState
         };
         return temp;
     };
@@ -127,30 +127,8 @@ var ButtonInit = function () {
 
     oInit.Init = function () {
         $("#btn_add").click(function () {
-            window.document.location=schoolMng.URL.add();
+            window.document.location = schoolMng.URL.add();
         });
-
-        //$("#btn_edit").click(function () {
-        //    var arrselections = $("#tb_departments").bootstrapTable('getSelections');
-        //    if (arrselections.length > 1) {
-        //        toastr.warning('只能选择一行进行编辑');
-
-        //        return;
-        //    }
-        //    if (arrselections.length <= 0) {
-        //        toastr.warning('请选择有效数据');
-
-        //        return;
-        //    }
-        //    $("#myModalLabel").text("编辑");
-        //    $("#txt_departmentname").val(arrselections[0].DEPARTMENT_NAME);
-        //    $("#txt_parentdepartment").val(arrselections[0].PARENT_ID);
-        //    $("#txt_departmentlevel").val(arrselections[0].DEPARTMENT_LEVEL);
-        //    $("#txt_statu").val(arrselections[0].STATUS);
-
-        //    postdata.DEPARTMENT_ID = arrselections[0].DEPARTMENT_ID;
-        //    $('#myModal').modal();
-        //});
 
         $("#btn_delete").click(function () {
             var arrselections = $("#tb_schools").bootstrapTable('getSelections');
@@ -159,53 +137,34 @@ var ButtonInit = function () {
                 return;
             }
             var select = new Array();
-            for (var i=0;i<arrselections.length;i++){
+            for (var i = 0; i < arrselections.length; i++) {
                 var selectSchool = new Object();
-                selectSchool.schId=arrselections[i].schId;
+                selectSchool.schId = arrselections[i].schId;
                 select.push(selectSchool);
             }
 
-            
-            $.ajax({
-                type: "post",
-                url: schoolMng.URL.delete(),
-                dataType:"json",
-                contentType:"application/json;charset=utf-8",
-                data: JSON.stringify(select),
-                success: function (data, status) {
-                    debugger;
-                    if (data.success) {
-                        toastr.success('删除成功');
-                        $("#tb_schools").bootstrapTable('refresh');
-                    }else {
-                        toastr.error(data.msg);
-                    }
+            Ewin.confirm({message: "确认要删除选择的学校吗？"}).on(function (e) {
+                if (!e) {
+                    return;
                 }
+                $.ajax({
+                    type: "post",
+                    url: schoolMng.URL.delete(),
+                    dataType: "json",
+                    contentType: "application/json;charset=utf-8",
+                    data: JSON.stringify(select),
+                    success: function (data, status) {
+                        debugger;
+                        if (data.success) {
+                            toastr.success('删除成功');
+                            $("#tb_schools").bootstrapTable('refresh');
+                        } else {
+                            toastr.error(data.msg);
+                        }
+                    }
 
+                });
             });
-           // Ewin.confirm({ message: "确认要删除选择的数据吗？" }).on(function (e) {
-           //     if (!e) {
-           //         return;
-           //     }
-           //     $.ajax({
-           //         type: "post",
-           //         url: "/school/schoolDelAction",
-           //         data: { "delList": JSON.stringify(arrselections) },
-           //         success: function (data, status) {
-           //             if (status == "success") {
-           //                 toastr.success('提交数据成功');
-           //                 $("#tb_schools").bootstrapTable('refresh');
-           //             }
-           //         },
-           //         error: function () {
-           //             toastr.error('Error');
-           //         },
-           //         complete: function () {
-           //
-           //         }
-           //
-           //     });
-           // });
         });
 
         $("#btn_disable").click(function () {
@@ -215,93 +174,75 @@ var ButtonInit = function () {
                 return;
             }
             var select = new Array();
-            for (var i=0;i<arrselections.length;i++){
+            for (var i = 0; i < arrselections.length; i++) {
                 var selectSchool = new Object();
-                selectSchool.schId=arrselections[i].schId;
-                selectSchool.state=arrselections[i].state;
+                selectSchool.schId = arrselections[i].schId;
+                selectSchool.state = 0;
+                select.push(selectSchool);
+            }
 
+            Ewin.confirm({message: "确认要禁用选择的学校吗？"}).on(function (e) {
+                if (!e) {
+                    return;
+                }
+                $.ajax({
+                    type: "post",
+                    url: schoolMng.URL.changeState(),
+                    dataType: "json",
+                    contentType: "application/json;charset=utf-8",
+                    data: JSON.stringify(select),
+                    success: function (data, status) {
+                        debugger;
+                        if (data.success) {
+                            toastr.success('禁用成功');
+                            $("#tb_schools").bootstrapTable('refresh');
+                        } else {
+                            toastr.error('禁用失败');
+                        }
+                    }
+                });
+            });
+        });
+
+
+        $("#btn_enable").click(function () {
+            var arrselections = $("#tb_schools").bootstrapTable('getSelections');
+            if (arrselections.length <= 0) {
+                toastr.warning('请选择至少一条数据');
+                return;
+            }
+            var select = new Array();
+            for (var i = 0; i < arrselections.length; i++) {
+                var selectSchool = new Object();
+                selectSchool.schId = arrselections[i].schId;
+                selectSchool.state = 1;
                 select.push(selectSchool);
             }
             $.ajax({
                 type: "post",
-                url: "/school/schoolDisableAction",
-                dataType:"json",
-                contentType:"application/json;charset=utf-8",
+                url: schoolMng.URL.changeState(),
+                dataType: "json",
+                contentType: "application/json;charset=utf-8",
                 data: JSON.stringify(select),
                 success: function (data, status) {
-                    debugger;
                     if (data.success) {
-                        toastr.success('禁用成功');
+                        toastr.success('启用成功');
                         $("#tb_schools").bootstrapTable('refresh');
-                    }else {
-                        toastr.error('禁用失败');
+                    } else {
+                        toastr.error('启用失败');
                     }
                 }
             });
-            // Ewin.confirm({ message: "确认要删除选择的数据吗？" }).on(function (e) {
-            //     if (!e) {
-            //         return;
-            //     }
-            //     $.ajax({
-            //         type: "post",
-            //         url: "/school/schoolDelAction",
-            //         data: { "delList": JSON.stringify(arrselections) },
-            //         success: function (data, status) {
-            //             if (status == "success") {
-            //                 toastr.success('提交数据成功');
-            //                 $("#tb_schools").bootstrapTable('refresh');
-            //             }
-            //         },
-            //         error: function () {
-            //             toastr.error('Error');
-            //         },
-            //         complete: function () {
-            //
-            //         }
-            //
-            //     });
-            // });
         });
-
-        //$("#btn_submit").click(function () {
-        //    postdata.DEPARTMENT_NAME = $("#txt_departmentname").val();
-        //    postdata.PARENT_ID = $("#txt_parentdepartment").val();
-        //    postdata.DEPARTMENT_LEVEL = $("#txt_departmentlevel").val();
-        //    postdata.STATUS = $("#txt_statu").val();
-        //    $.ajax({
-        //        type: "post",
-        //        url: "/Home/GetEdit",
-        //        data: { "": JSON.stringify(postdata) },
-        //        success: function (data, status) {
-        //            if (status == "success") {
-        //                toastr.success('提交数据成功');
-        //                $("#tb_departments").bootstrapTable('refresh');
-        //            }
-        //        },
-        //        error: function () {
-        //            toastr.error('Error');
-        //        },
-        //        complete: function () {
-
-        //        }
-
-        //    });
-        //});
-
-        //$("#btn_query").click(function () {
-        //    $("#tb_departments").bootstrapTable('refresh');
-        //});
     };
-
     return oInit;
 };
 
 
-
-
 $(function () {
-    //debugger;
     schoolMng.init();
     //初始化消息框位置
     myToastr.init();
+    //初始化模态框位置
+    myDialog.init();
 });
